+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
+## v3.0.0 - April 9, 2017
+
+- Position drag handles relative to editor, not document
+- Simplify toolbar and use floats
+- Improved styling
+
 ## v2.0.0 - April 3, 2017
 
-- Initial version
+- Added toolbar for alignment
+- Modularized
+- Fix resize bug
 
 ## v1.0.0 - March 25, 2017